Win32: Improve clipboard contention issue
This is primarily a workaround for a GLFW application reading and/or
writing to the clipboard in rapid succession and catching up with the
Windows 10 Clipboard History, which also has to contend for the lock.

As a side-effect it will also improve our interaction with other
programs, excluding buggy ones that leave it open for a long time.